\name{Taipackage}
\alias{inclass}
\title{Plot interactive image intensity graphs}
\usage{
find_ratio()
fig_ratio()
full_analysis()
}
\description{
This package works on Image J image intensity measurement's csv file,
it gets the ratio of the Integrated intensity and raw intensity and makes an interactive dot plot for the value to the area of the cell.
}
\examples{
find_ratio()
}
